+++ src/map_cache.h
@@ -53,9 +53,9 @@ namespace SlavGPS {
 		static void init(void);
 		static void uninit(void);
 
-		static void add(const QPixmap & pixmap, MapCacheItemExtra & extra, TileInfo * tile_info, MapTypeID map_type, uint8_t alpha, double xshrinkfactor, double yshrinkfactor, const QString & file_name);
-		static QPixmap get_pixmap(TileInfo * tile_info, MapTypeID map_type, uint8_t alpha, double xshrinkfactor, double yshrinkfactor, const QString & file_name);
-		static MapCacheItemExtra get_extra(TileInfo * tile_info, MapTypeID map_type, uint8_t alpha, double xshrinkfactor, double yshrinkfactor, const QString & file_name);
+		static void add(const QPixmap & pixmap, MapCacheItemExtra & extra, TileInfo * tile_info, MapTypeID map_type, int alpha, double xshrinkfactor, double yshrinkfactor, const QString & file_name);
+		static QPixmap get_pixmap(TileInfo * tile_info, MapTypeID map_type, int alpha, double xshrinkfactor, double yshrinkfactor, const QString & file_name);
+		static MapCacheItemExtra get_extra(TileInfo * tile_info, MapTypeID map_type, int alpha, double xshrinkfactor, double yshrinkfactor, const QString & file_name);
 
 		static size_t get_size(void);
 		static int get_count(void);
